Fix math calls to be float versions
Avoids upcasting intermediate results.

+++ code/add_numbers/openmp/add_numbers.c
@@ -22,10 +22,10 @@ void add_numbers(int n_numbers, float *numbers) {
 
   /* do the actual calculation */
   for (int i = 0; i < n_numbers; i++) {
-    result += fabs(log(pow(fabs(numbers[i]), 2.1))) +
-              log(pow(fabs(numbers[i]), 1.9)) +
-              log(pow(fabs(numbers[i]), -1.97)) +
-              fabs(log(pow(fabs(numbers[i]), -1.005)));
+    result += fabsf(logf(powf(fabsf(numbers[i]), 2.1))) +
+              logf(powf(fabsf(numbers[i]), 1.9)) +
+              logf(powf(fabsf(numbers[i]), -1.97)) +
+              fabsf(logf(powf(fabsf(numbers[i]), -1.005)));
   }
 
   /* time the calculation */
